pyproject.toml
[project]
name = "stereotype-bench"
version = "0.1.0"
description = "Measure-agnostic LLM stereotype benchmark engine."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [{ name = "Jake Azcona" }]
keywords = ["llm", "benchmark", "stereotype", "bias", "evaluation"]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]
dependencies = [
  "typer>=0.12",
  "httpx>=0.27",
  "pydantic>=2.9",
  "pyyaml>=6.0",
  "rich>=13.7",
  "matplotlib>=3.9",
  "pandas>=2.2",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3",
  "pytest-cov>=5.0",
  "ruff>=0.7",
]

[project.scripts]
stereotype-bench = "stereotype_bench.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/stereotype_bench"]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra --strict-markers"
